migrate for new sveltekit routes setup
