+++ site/package.json
@@ -36,31 +36,31 @@
 		"tauri": "tauri"
 	},
 	"devDependencies": {
-		"@playwright/test": "^1.24.1",
+		"@playwright/test": "^1.25.0",
 		"@sveltejs/adapter-auto": "^1.0.0-next.64",
-		"@sveltejs/kit": "^1.0.0-next.396",
+		"@sveltejs/kit": "^1.0.0-next.427",
 		"@tauri-apps/cli": "^1.0.5",
 		"@tweakpane/core": "^1.1.0",
 		"@types/three": "^0.141.0",
 		"@types/uuid": "^8.3.4",
-		"@typescript-eslint/eslint-plugin": "^5.31.0",
-		"@typescript-eslint/parser": "^5.31.0",
+		"@typescript-eslint/eslint-plugin": "^5.33.1",
+		"@typescript-eslint/parser": "^5.33.1",
 		"eslint": "^7.32.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-svelte3": "^3.4.1",
 		"prettier": "^2.7.1",
 		"prettier-plugin-svelte": "^2.7.0",
-		"prisma": "^4.1.0",
+		"prisma": "^4.2.1",
 		"svelte": "^3.49.0",
-		"svelte-check": "^2.8.0",
+		"svelte-check": "^2.8.1",
 		"svelte-preprocess": "^4.10.7",
 		"three": "^0.141.0",
 		"tslib": "^2.4.0",
 		"typescript": "~4.6.2",
-		"vite": "^3.0.3"
+		"vite": "^3.0.9"
 	},
 	"dependencies": {
-		"@prisma/client": "^4.1.0",
+		"@prisma/client": "^4.2.1",
 		"ansi-regex": "^6.0.1",
 		"glob-parent": "^6.0.2",
 		"json-schema": "^0.4.0",
