+++ site/src/routes/+page.ts
@@ -1,0 +1,8 @@
+import { defaultRecipe } from './_recipe';
+
+/** @type {import('./index').PageLoad} */
+export async function load() {
+	return {
+		recipe: defaultRecipe
+	};
+}
